+++ role.operator.js
@@ -16,24 +16,55 @@ PowerCreep.prototype.runOperator = function (plant_data) {
 
     // if we don't have a task currently assigned
     if (this.memory.task == null) {
-        if (plant_data.factory_state == STATES.CLEANING) {
-            // TODO: find a resource to cleanup from the factory
-            // TODO: assign a move resource task
-        }else if (plant_data.labs_state == STATES.CLEANING){
-            // TODO: find a resource to cleanup from the labs
-            // TODO: assign a move resource task
-        }else if (plant_data.factory_state == STATES.FINISHED){
-            // TODO: find a resource to cleanup from the factory
-            // TODO: assign a move resource task
-        }else if (plant_data.labs_state == STATES.FINISHED){
-            // TODO: find a resource to cleanup from the labs
-            // TODO: assign a move resource task
+        if (plant_data.factory_state == STATES.CLEANING || plant_data.factory_state == STATES.FINISHED) {
+            let factory = Game.getObjectById(plant_data.factory_id);
+            if (factory != null && this.room.storage != undefined) {
+                for (let resource in factory.store) {
+                    if (factory.store[resource] > 0) {
+                        this.memory.task = new MoveResourceTask(this.room.name, factory, this.room.storage, resource, factory.store[resource]);
+                        break;
+                    }
+                }
+            }
+        }else if (plant_data.labs_state == STATES.CLEANING || plant_data.labs_state == STATES.FINISHED){
+            let labs = [plant_data.input_lab_1_id, plant_data.input_lab_2_id, plant_data.output_lab_id];
+            for (let i = 0; i < labs.length; i++) {
+                let lab = Game.getObjectById(labs[i]);
+                if (lab != null && this.room.storage != undefined) {
+                    for (let resource in lab.store) {
+                        if (lab.store[resource] > 0) {
+                            this.memory.task = new MoveResourceTask(this.room.name, lab, this.room.storage, resource, lab.store[resource]);
+                            break;
+                        }
+                    }
+                    if (this.memory.task != null) {
+                        break;
+                    }
+                }
+            }
         }else if (plant_data.factory_state == STATES.LOADING){
-            // TODO: find a resource to load for the factory
-            // TODO: assign a move resource task
+            let production = plant_data.current_production;
+            let factory = Game.getObjectById(plant_data.factory_id);
+
+            if (factory != null && this.room.storage != undefined) {
+                for (let resource in production.inputs) {
+                    if (factory.store[resource] == undefined || factory.store[resource] < production.inputs[resource]) {
+                        this.memory.task = new MoveResourceTask(this.room.name, this.room.storage, factory, resource, (production.inputs[resource] - factory.store[resource]));
+                        break;
+                    }
+                }
+            }
         }else if (plant_data.labs_state == STATES.LOADING){
-            // TODO: find a resource to load for the labs
-            // TODO: assign a move resource task
+            let reaction = plant_data.current_reaction;
+            let input_lab_1 = Game.getObjectById(plant_data.input_lab_1_id);
+            let input_lab_2 = Game.getObjectById(plant_data.input_lab_2_id);
+            if (input_lab_1 != null && input_lab_2 != null && this.room.storage != undefined) {
+                if (input_lab_1.store[reaction.input_1] == undefined || input_lab_1.store[reaction.input_1] < reaction.amount) {
+                    this.memory.task = new MoveResourceTask(this.room.name, this.room.storage, input_lab_1, reaction.input_1, (reaction.amount - input_lab_1.store[reaction.input_1]));
+                }else if (input_lab_2.store[reaction.input_2] == undefined || input_lab_2.store[reaction.input_2] < reaction.amount) {
+                    this.memory.task = new MoveResourceTask(this.room.name, this.room.storage, input_lab_2, reaction.input_2, (reaction.amount - input_lab_2.store[reaction.input_2]));
+                }
+            }
         }else{
             this.memory.task = new IdleTask(this.room.name);
         }