+++ role.queen.js
@@ -38,41 +38,79 @@ Creep.prototype.runQueen = function () {
     // if we don't have a task currently assigned
     if (this.memory.task == null) {
         // if we are out of energy
-        if (this.store[RESOURCE_ENERGY] == 0) {
-            // grab the storage in the room
-            let target = this.room.storage;
-
-            // TODO: if the storage has any ingredients in it and this isn't a capitol
-                // TODO: assign a task to move ingredients from the storage to the terminal
-            // TODO: if the terminal has any ingredients in it and this is a capitol
-                // TODO: assign a task to move ingredients from the terminal to the storage
+        if (this.store.getUsedCapacity() == 0) {
+            let storage = this.room.storage;
+            let terminal = this.room.terminal;
 
             // if the storage isn't built or is empty
-            if (target == undefined || target.store[RESOURCE_ENERGY] == 0) {
+            if (storage == undefined || storage.store.getUsedCapacity() == 0) {
                 // assign a new task
                 this.memory.task = new IdleTask(this.memory.room_name, 10);
                 // announce the new task
                 this.announceTask();
             }else{
-                // assign a new task
-                this.memory.task = new GatherTask(target, RESOURCE_ENERGY);
-                // announce the new task
-                this.announceTask();
+                if (this.room.name == Memory.capitol_room_name) {
+                    for (let resource in storage.store) {
+                        if (storage.store[resource] > 0 && FINAL_PRODUCTS.includes(resource) && terminal != undefined) {
+                            this.memory.task = new MoveResourceTask(this.room.name, storage, terminal, resource, storage.store[resource]);
+                        }
+                    }
+                    if (this.memory.task != null && terminal != undefined) {
+                        for (let resource in terminal.store) {
+                            if (terminal.store[resource] > 0 && INGREDIENTS.includes(resource)) {
+                                this.memory.task = new MoveResourceTask(this.room.name, terminal, storage, resource, storage.store[resource]);
+                            }
+                        }
+                    }
+                    if (this.memory.task != null) {
+                        this.memory.task = new IdleTask(this.room.name);
+                    }
+                }else{
+                    if (terminal != undefined) {
+                        for (let resource in storage.store) {
+                            if (storage.store[resource] > 0 && resource != RESOURCE_ENERGY) {
+                                this.memory.task = new MoveResourceTask(this.room.name, storage, terminal, resource, storage.store[resource]);
+                            }
+                        }
+                    }else{
+                        this.memory.task = new IdleTask(this.room.name);
+                    }
+                }
+
+                if (this.memory.task != null) {
+                    this.memory.task = new GatherTask(storage,  RESOURCE_ENERGY);
+                }
             }
         }else{
-            // find a new dump target fit for a queen
-            let target = this.getQueenDumpTarget();
-            // if a new target was found
-            if (target != null) {
-                // assign a new task
-                this.memory.task = new DepositTask(target, RESOURCE_ENERGY);
-                // announce the new task
-                this.announceTask();
+            if (this.store.getUsedCapacity() == this.store[RESOURCE_ENERGY]) {
+                // find a new dump target fit for a queen
+                let target = this.getQueenDumpTarget();
+                // if a new target was found
+                if (target != null) {
+                    // assign a new task
+                    this.memory.task = new DepositTask(target, RESOURCE_ENERGY);
+                    // announce the new task
+                    this.announceTask();
+                }else{
+                    // assign a new task
+                    this.memory.task = new IdleTask(this.memory.room_name, 10);
+                    // announce the new task
+                    this.announceTask();
+                }
             }else{
-                // assign a new task
-                this.memory.task = new IdleTask(this.memory.room_name, 10);
-                // announce the new task
-                this.announceTask();
+                let storage = this.room.storage;
+                if (storage != undefined) {
+                    for (let resource in this.store) {
+                        if (this.store[resource] > 0) {
+                            this.memory.task = new DepositTask(this.room.storage, resource);
+                        }
+                    }
+                    if (this.memory.task != null) {
+                        this.memory.task = new IdleTask(this.memory.room_name, 10);
+                    }
+                }else{
+                    this.memory.task = new IdleTask(this.memory.room_name, 10);
+                }
             }
         }
     }